+++ R/npv.R
@@ -13,27 +13,24 @@
 #' @export npv
 #' @author Philippe Cote
 #' @examples
-#' \dontrun{
-#' us.df <- ir_df_us(quandlkey = quandlkey, ir.sens = 0.01)
 #' npv(
 #'   init.cost = -375, C = 50, cf.freq = .5, TV = 250, T2M = 2,
-#'   disc.factors = us.df, BreakEven = TRUE, BE.yield = .0399
+#'   disc.factors = RTL::usSwapCurves, BreakEven = FALSE, BE.yield = .0399
 #' )$npv
 #' npv(
 #'   init.cost = -375, C = 50, cf.freq = .5, TV = 250, T2M = 2,
-#'   disc.factors = us.df, BreakEven = TRUE, BE.yield = .0399
+#'   disc.factors = RTL::usSwapCurves, BreakEven = FALSE, BE.yield = .0399
 #' )$df
-#' }
 npv <- function(init.cost = -375, C = 50, cf.freq = .25, TV = 250, T2M = 2, disc.factors = us.df, BreakEven = FALSE, BE.yield = .01) {
   if (BreakEven == TRUE) {
     disc.factors$yield <- BE.yield
-    disc.factors <- disc.factors %>% dplyr::mutate(discountfactor = exp(-yield * maturity))
+    disc.factors$discounts <- exp(-BE.yield * disc.factors$times)
   }
   df <- dplyr::tibble(t = seq(from = 0, to = T2M, by = cf.freq), cf = C) %>%
     dplyr::mutate(
       cf = replace(cf, t == 0, init.cost),
       cf = replace(cf, t == T2M, TV),
-      df = stats::spline(x = disc.factors$maturity, y = disc.factors$discountfactor, xout = t)$y,
+      df = stats::spline(x = disc.factors$times, y = disc.factors$discounts, xout = t)$y,
       pv = cf * df
     )
   x <- list(df = df, npv = sum(df$pv))