+++ R/tradeStats.R
@@ -7,11 +7,7 @@
 #' @author Philippe Cote
 #' @examples
 #' library(PerformanceAnalytics)
-#' x <- tidyquant::tq_get("SPY") %>% dplyr::mutate(ret = log(adjusted / dplyr::lag(adjusted)))
-#' x <- x %>%
-#'   stats::na.omit() %>%
-#'   dplyr::select(date, ret)
-#' tradeStats(x = x, Rf = 0)
+#' tradeStats(x = RTL::spy, Rf = 0)
 tradeStats <- function(x, Rf = 0) {
   # if (!requireNamespace("PerformanceAnalytics", quietly = TRUE)) {
   #   stop("Package \"PerformanceAnalytics\" needed for this function to work. Please install it.",